Move DRK Action Logic into its own file
